make shapes library not depend on core but only on base
